Restructure webserver code
