+++ autosubliminal/server/web/log.py
@@ -1,0 +1,38 @@
+# coding=utf-8
+
+import os
+import re
+
+import cherrypy
+
+import autosubliminal
+from autosubliminal import utils
+from autosubliminal.server.web import redirect
+from autosubliminal.templates.page import PageTemplate
+
+
+class Log(object):
+    def __init__(self):
+        self.template_file = '/log/log.mako'
+        pass
+
+    @cherrypy.expose
+    def index(self):
+        redirect('/log/viewLog')
+
+    @cherrypy.expose(alias='viewLog')
+    def view_log(self, loglevel='all', lognum=None):
+        content = utils.display_logfile(loglevel, lognum)
+        return PageTemplate(filename=self.template_file).render(loglevel=loglevel, lognum=lognum, content=content)
+
+    @cherrypy.expose(alias='clearLog')
+    def clear_log(self):
+        # Clear log file (open it in write mode and pass)
+        with open(autosubliminal.LOGFILE, 'w'):
+            pass
+        # Remove possible backup log files
+        for f in [f for f in os.listdir('.') if os.path.isfile(f) and re.match(autosubliminal.LOGFILE + '.', f)]:
+            os.remove(f)
+        # Return to default log view
+        content = utils.display_logfile()
+        return PageTemplate(filename=self.template_file).render(loglevel='all', lognum=None, content=content)
